Review issues: add KDoc to the FunctionProvider::merge(), FunctionProvider::mergeSelect() functions; rename MergeBaseStatement to MergeStatement
