+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/statements/MergeSelectStatement.kt
@@ -10,7 +10,7 @@ import org.jetbrains.exposed.sql.Transaction
  * or deletions.
  *
  * Here is only the part specific for the Query as a source implementation.
- * Look into [MergeBaseStatement] to find the base implementation of that command.
+ * Look into [MergeStatement] to find the base implementation of that command.
  *
  * @property dest The destination [Table] where records will be merged into.
  * @property selectQuery The source [QueryAlias] from which records are taken to compare with `dest`.
@@ -20,7 +20,7 @@ class MergeSelectStatement(
     dest: Table,
     private val selectQuery: QueryAlias,
     val on: Op<Boolean>
-) : MergeBaseStatement(dest) {
+) : MergeStatement(dest) {
     override fun prepareSQL(transaction: Transaction, prepared: Boolean): String {
         return transaction.db.dialect.functionProvider.mergeSelect(table, selectQuery, transaction, whenBranches, on, prepared)
     }
