+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/statements/MergeTableStatement.kt
@@ -9,7 +9,7 @@ import org.jetbrains.exposed.sql.Transaction
  * or deletions.
  *
  * Here is only the part specific for the Table as a source implementation.
- * Look into [MergeBaseStatement] to find the base implementation of that command.
+ * Look into [MergeStatement] to find the base implementation of that command.
  *
  * @property dest The destination [Table] where records will be merged into.
  * @property source The source [Table] from which records are taken to compare with `dest`.
@@ -19,7 +19,7 @@ open class MergeTableStatement(
     dest: Table,
     private val source: Table,
     private val on: Op<Boolean>?
-) : MergeBaseStatement(dest) {
+) : MergeStatement(dest) {
     override fun prepareSQL(transaction: Transaction, prepared: Boolean): String {
         val result = transaction.db.dialect.functionProvider.merge(table, source, transaction, whenBranches, on)
         return result
